RDoc-2475 Client API > Session > Querying > Count query results - Part 1
