+++ Documentation/5.2/Raven.Documentation.Pages/client-api/session/querying/how-to-query.js.markdown
@@ -265,21 +265,21 @@ Available methods for the session's [query](../../../client-api/session/querying
 - closeSubclause
 - containsAll
 - containsAny
-- count
-- longCount
-- countLazily
+- [count](../../../client-api/session/querying/how-to-count-query-results)
+- [countLazily](../../../client-api/session/querying/how-to-perform-queries-lazily#lazy-count-query)
 - distinct
 - first
 - firstOrNull
 - fuzzy
+- getIndexQuery
 - getQueryResult
 - [groupBy](../../../client-api/session/querying/how-to-perform-group-by-query)
 - [highlight](../../../client-api/session/querying/how-to-use-highlighting)
 - include
 - [includeExplanations](../../../client-api/session/querying/debugging/include-explanations)
-- getIndexQuery
 - [intersect](../../../client-api/session/querying/how-to-use-intersect)
 - lazily
+- [longCount](../../../client-api/session/querying/how-to-count-query-results)
 - [moreLikeThis](../../../client-api/session/querying/how-to-use-morelikethis)
 - negateNext
 - [noCaching](../../../client-api/session/querying/how-to-customize-query#nocaching)
