+++ Documentation/6.0/Raven.Documentation.Pages/client-api/session/querying/.docs.json
@@ -94,7 +94,7 @@
     },
     {
         "Path": "how-to-count-query-results.markdown",
-        "Name": "How to Count Query Results",
+        "Name": "Count Query Results",
         "DiscussionId": "77cec3b4-0001-4a30-81a7-e3182f4150bf",
         "Mappings": []
     },
